led_panel_driver: fix incorrect point_to_index
The typedef for point_to_index expects driver as the first parameter.
For some reason -Werror was not turned on even though ESP says it is
on by default. Fix this function, add explicit error checks to cmake
definition to avoid this issue in the future. Also make points and
sizes completely consistent and unsigned throughout.

+++ main/renderer/surface.h
@@ -75,8 +75,8 @@ struct surface_t {
  * @brief Surface configuration
  */
 typedef struct {
-    int width;  /*!< Width of the surface in pixels*/
-    int height; /*!< Height of the surface in pixels*/
+    uint16_t width;  /*!< Width of the surface in pixels*/
+    uint16_t height; /*!< Height of the surface in pixels*/
 } surface_config_t;
 
 /**
